+++ waveframe_guard/__init__.py
@@ -3,7 +3,7 @@ from .context import install_guard, resolve_actor
 from .execute import GovernanceError, execute
 from .contracts import load_contract
 from .result import GovernedExecutionResult
-from .runtime import GovernedRuntime, evaluate_admissibility
+from .runtime import GuardRuntime, GovernedRuntime, evaluate_admissibility
 from .schemas import (
     GOVERNED_EXECUTION_EVENT_V1,
     GOVERNED_EXECUTION_RESULT_V1,
@@ -21,6 +21,7 @@ __all__ = [
     "load_contract",
     "GovernedExecutionResult",
     "GovernedRuntime",
+    "GuardRuntime",
     "evaluate_admissibility",
     "GOVERNED_EXECUTION_EVENT_V1",
     "GOVERNED_EXECUTION_RESULT_V1",